Required settings config for tests
