+++ tests/api/metrics/conftest.py
@@ -1,17 +1,20 @@
 from functools import lru_cache
 
 import pytest
-from fastapi.testclient import TestClient
 from prometheus_client import REGISTRY
 
 from gh_actions_exporter.config import Relabel, Settings
-from gh_actions_exporter.main import app, get_settings, metrics
+from gh_actions_exporter.main import get_settings, metrics
 from gh_actions_exporter.metrics import Metrics
 
 
 @lru_cache()
 def job_relabel_config():
     return Settings(
+        check_runs_enabled=False,
+        github_app_id=42,
+        github_app_installation_id=42,
+        github_app_private_key="private key",
         job_relabelling=[
             Relabel(
                 label="cloud", values=["mycloud"], type="name", default="github-hosted"
@@ -29,6 +32,19 @@ def job_relabel_config():
 def relabel_metrics():
     return Metrics(job_relabel_config())
 
+@lru_cache()
+def default_settings():
+    return Settings(
+        check_runs_enabled=False,
+        github_app_id=42,
+        github_app_installation_id=42,
+        github_app_private_key="private key"
+    )
+
+@lru_cache()
+def default_metrics():
+    return Metrics(default_settings())
+
 
 def unregister_metrics():
     print(f"Unregistering {REGISTRY._collector_to_names}")
@@ -37,92 +53,17 @@ def unregister_metrics():
             REGISTRY.unregister(collector)
 
 
+# Retrieving the existent fastapi fixture from the main conftest.py
+# and overriding the dependency overrides.
+@pytest.fixture(scope="function")
+def fastapp(fastapp):
+    fastapp.dependency_overrides[get_settings] = default_settings
+    fastapp.dependency_overrides[metrics] = default_metrics
+    return fastapp
+
+
 @pytest.fixture(scope="function")
 def override_job_config(fastapp):
     unregister_metrics()
     fastapp.dependency_overrides[get_settings] = job_relabel_config
     fastapp.dependency_overrides[metrics] = relabel_metrics
-
-
-@pytest.fixture(scope="function", autouse=True)
-def fastapp():
-    fastapp = app
-    return fastapp
-
-
-@pytest.fixture(scope="function")
-def client(fastapp):
-    client = TestClient(fastapp)
-    return client
-
-
-@pytest.fixture(scope="function", autouse=True)
-def destroy_client(client):
-    client.delete("/clear")
-
-
-@pytest.fixture
-def workflow_run():
-    webhook = {
-        "workflow_run": {
-            "id": 4863423668,
-            "name": "test",
-            "head_branch": "feature",
-            "head_sha": "546c86cb55726fd1e647b9a886c0c5ee63ca0718",
-            "run_number": 42,
-            "run_attempt": 1,
-            "event": "push",
-            "status": "completed",
-            "conclusion": "success",
-            "workflow_id": 13826527,
-            "created_at": "2021-11-16T17:52:47Z",
-            "run_started_at": "2021-11-16T17:52:47Z",
-            "updated_at": "2021-11-16T17:53:32Z",
-        },
-        "repository": {
-            "name": "runners-test",
-            "full_name": "scalanga-devl/runners-test",
-            "visibility": "private",
-        },
-        "organization": {"login": "scalanga-devl"},
-    }
-    return webhook
-
-
-@pytest.fixture
-def workflow_job():
-    webhook = {
-        "workflow_job": {
-            "id": 4355026428,
-            "run_id": 1468134741,
-            "run_url": "https://api.github.com/repos/-devl/fake",
-            "run_attempt": 1,
-            "node_id": "CR_kwDOFqdmms8AAAABA5Rt_A",
-            "head_sha": "dfcc88973035bece77c0312f3f29e4e123e1f952",
-            "url": "https://api.github.com/repos/scalanga-devl/fake",
-            "html_url": "https://github.com/scalanga-devl/fake",
-            "status": "in_progress",
-            "conclusion": None,
-            "started_at": "2021-11-29T14:46:57Z",
-            "completed_at": None,
-            "name": "greet (tata)",
-            "runner_name": "GitHub Actions 3",
-            "steps": [
-                {
-                    "name": "Set up job",
-                    "status": "completed",
-                    "conclusion": "success",
-                    "number": 1,
-                    "started_at": "2021-11-29T14:50:57Z",
-                    "completed_at": None,
-                }
-            ],
-            "labels": ["ubuntu-latest"],
-        },
-        "repository": {
-            "name": "test-runner-operator",
-            "full_name": "scalanga-devl/test-runner-operator",
-            "visibility": "private",
-        },
-    }
-    return webhook
